+++ changelog.md
@@ -36,6 +36,7 @@ This version is not yet released. If you are reading this on the website, then t
   - Sorting is very common
 - Remove previously deprecated function strands
 ### Interpreter
+- Add the `uiua find` command, which finds Uiua code that matches the given unformatted text
 - Some optimizations
   - [`rows ≡`](https://uiua.org/docs/rows) [`on ⟜`](https://uiua.org/docs/on)/[`by ⊸`](https://uiua.org/docs/by) [`rand ⚂`](https://uiua.org/docs/rand)/`constant`
 - Tweak the formatter to reduce excess vertical space